requirements.txt
fastapi==0.70.1
pydantic==1.8.2
uvicorn==0.16.0
transformers==4.15.0
openai==0.23.1
python-dotenv==0.20.0
gunicorn==20.1.0